Bump dependencies, `26.0.0-beta2` :arrow_right: `26.+`
- even though Android Studio warns us that we shouldn't "float"
dependencies (i.e. `26.+` is bad), the default behavior on a greenfield
Android project is to float dependencies (i.e. it's the canonical
behavior). So we do as Android Studio does, not as it says.
- Also we do a slight bump to espresso, 2.2.1 -> 2.2.2. There is a 3.0.1
out there, but even though the documentation indicates it's available,
it doesn't appear to work in the stable version of Android Studio
(2.3.3). It appears to work in the beta Android Studio 3, but switching
to a beta Android Studio seems to invite more problems.