New Doc: Diarizer Benchmarks
I want to rip this document out of my private Google folder and share it with
everyone who is interested.  It needs formatting to convert it to true
markdown. I think this is a good template of my progress, and I hope that by
publishing it & investing time to keep it up-to-date, it will help focus my
research on what really is the best diarizer for BlabberTabber.